// src/services/features/users/types.ts
export interface IUsersData {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}

export interface IUsersDataExtended extends IUsersData {
  img: string;
  workspace?: string;
  privacy?: string;
  security?: string;
  isArchived?: boolean;
}